CLN: Move AggregatedData to own file
